chore(desktop): drop 22 unused UI components + orphan test mock
Phase 1 of the Prism UI refactor (see docs/superpowers/specs/
2026-04-26-prism-ui-component-library-design.md). Every file removed
here had zero active alias-form consumers per `grep -rln "from \"@/components/ui/<n>\""`.

Removed components: breadcrumbs, collapsible, context-menu, dropdown-menu,
dropzone, field, link, list-box, menu, numberfield, popover, radio-group,
scroll-area, searchfield, separator, sheet, slider, switch, table, tag-group,
textfield, toggle-group.

Three of the deleted files (field, list-box, popover) were still imported
via relative paths by surviving siblings. Inlined the minimum primitives
the surviving consumers (checkbox.tsx, progress.tsx, select.tsx) used from
those modules — kept the diff to direct RAC imports / inlined primitives
only. No new behavior; structural unblock so Phase 1's full delete set
can stand.

Also removed src/test/radix-radio-group-mock.tsx and the matching
vi.mock("@/components/ui/radio-group", …) line in src/test/setup.ts —
they existed for a consumer that has since been removed.

Six of the deleted files were direct frontend-no-axios-no-radix-no-danger.md
violations (Radix primitives: collapsible, context-menu, dropdown-menu,
scroll-area, sheet, toggle-group); eight more imported @radix-ui/react-icons
(breadcrumbs, list-box, menu, numberfield, radio-group, searchfield, table,
tag-group); the remaining eight were pure Jolly with no consumer.
