--- crates/rdlp-desktop/src/components/ui/textfield.tsx
@@ -1,95 +1,0 @@
-"use client"
-
-import {
-  Input as AriaInput,
-  InputProps as AriaInputProps,
-  TextArea as AriaTextArea,
-  TextAreaProps as AriaTextAreaProps,
-  TextField as AriaTextField,
-  TextFieldProps as AriaTextFieldProps,
-  ValidationResult as AriaValidationResult,
-  composeRenderProps,
-  Text,
-} from "react-aria-components"
-
-import { cn } from "@/lib/utils"
-
-import { FieldError, Label } from "./field"
-
-const TextField = AriaTextField
-
-const Input = ({ className, ...props }: AriaInputProps) => {
-  return (
-    <AriaInput
-      className={composeRenderProps(className, (className) =>
-        cn(
-          "flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 text-sm shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground",
-          /* Disabled */
-          "data-[disabled]:cursor-not-allowed data-[disabled]:opacity-50",
-          /* Focused */
-          "data-[focused]:outline-none data-[focused]:ring-1 data-[focused]:ring-ring",
-          /* Resets */
-          "focus-visible:outline-none",
-          className
-        )
-      )}
-      {...props}
-    />
-  )
-}
-
-const TextArea = ({ className, ...props }: AriaTextAreaProps) => {
-  return (
-    <AriaTextArea
-      className={composeRenderProps(className, (className) =>
-        cn(
-          "flex min-h-[80px] w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground",
-          /* Focused */
-          "data-[focused]:outline-none data-[focused]:ring-2 data-[focused]:ring-ring data-[focused]:ring-offset-2",
-          /* Disabled */
-          "data-[disabled]:cursor-not-allowed data-[disabled]:opacity-50",
-          /* Resets */
-          "focus-visible:outline-none",
-          className
-        )
-      )}
-      {...props}
-    />
-  )
-}
-
-interface JollyTextFieldProps extends AriaTextFieldProps {
-  label?: string
-  description?: string
-  errorMessage?: string | ((validation: AriaValidationResult) => string)
-  textArea?: boolean
-}
-
-function JollyTextField({
-  label,
-  description,
-  errorMessage,
-  textArea,
-  className,
-  ...props
-}: JollyTextFieldProps) {
-  return (
-    <TextField
-      className={composeRenderProps(className, (className) =>
-        cn("group flex flex-col gap-2", className)
-      )}
-      {...props}
-    >
-      <Label>{label}</Label>
-      {textArea ? <TextArea /> : <Input />}
-      {description && (
-        <Text className="text-sm text-muted-foreground" slot="description">
-          {description}
-        </Text>
-      )}
-      <FieldError>{errorMessage}</FieldError>
-    </TextField>
-  )
-}
-
-export { Input, TextField, JollyTextField, TextArea }
